requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.2
django-cors-headers==4.4.0
django-decouple==2.1
django-extensions==3.2.3
django-filter==23.3
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-extensions==0.7.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
google-api-core==2.19.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gunicorn==20.1.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
Pillow==11.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg[binary]==3.2.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
validate-docbr==1.10.0
wheel==0.43.0
